refactor(onboard): extract Docker GPU supervisor-reconnect debounce module + document env
